requirements_dev.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
bidict==0.22.1
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
distro==1.9.0
Faker==22.2.0
fastapi==0.109.0
filelock==3.13.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
openai==1.7.2
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.11.0
PyYAML==6.0.1
requests==2.31.0
ruff==0.1.14
sentry-sdk==1.39.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
virtualenv==20.25.0
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
